Update Id to Integer

+++ backend/src/main/java/com/activekids/web/repositories/AddressRepository.java
@@ -3,5 +3,5 @@ package com.activekids.web.repositories;
 import com.activekids.web.model.Address;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AddressRepository extends CrudRepository<Address, Long> {
+public interface AddressRepository extends CrudRepository<Address, Integer> {
 }
